+++ config/projection_test.go
@@ -20,7 +20,7 @@ var _ = Describe("type ProjectionConfig", func() {
 		BeforeEach(func() {
 			handler = &fixtures.ProjectionMessageHandler{
 				ConfigureFunc: func(c dogma.ProjectionConfigurer) {
-					c.Name("<name>")
+					c.Identity("<name>", "<key>")
 					c.ConsumesEventType(fixtures.MessageA{})
 					c.ConsumesEventType(fixtures.MessageB{})
 				},
@@ -36,13 +36,15 @@ var _ = Describe("type ProjectionConfig", func() {
 				Expect(err).ShouldNot(HaveOccurred())
 			})
 
-			It("the handler name is set", func() {
-				Expect(cfg.HandlerName).To(Equal("<name>"))
-			})
-
 			Describe("func Name()", func() {
 				It("returns the handler name", func() {
 					Expect(cfg.Name()).To(Equal("<name>"))
+				})
+			})
+
+			Describe("func Key()", func() {
+				It("returns the handler key", func() {
+					Expect(cfg.Key()).To(Equal("<key>"))
 				})
 			})
 
@@ -91,26 +93,34 @@ var _ = Describe("type ProjectionConfig", func() {
 				nil,
 			),
 			Entry(
-				"when the handler does not configure a name",
-				`*fixtures.ProjectionMessageHandler.Configure() did not call ProjectionConfigurer.Name()`,
+				"when the handler does not configure an identity",
+				`*fixtures.ProjectionMessageHandler.Configure() did not call ProjectionConfigurer.Identity()`,
 				func(c dogma.ProjectionConfigurer) {
 					c.ConsumesEventType(fixtures.MessageA{})
 				},
 			),
 			Entry(
-				"when the handler configures multiple names",
-				`*fixtures.ProjectionMessageHandler.Configure() has already called ProjectionConfigurer.Name("<name>")`,
+				"when the handler configures multiple identities",
+				`*fixtures.ProjectionMessageHandler.Configure() has already called ProjectionConfigurer.Identity("<name>", "<key>")`,
 				func(c dogma.ProjectionConfigurer) {
-					c.Name("<name>")
-					c.Name("<other>")
+					c.Identity("<name>", "<key>")
+					c.Identity("<other>", "<key>")
 					c.ConsumesEventType(fixtures.MessageA{})
 				},
 			),
 			Entry(
 				"when the handler configures an invalid name",
-				`*fixtures.ProjectionMessageHandler.Configure() called ProjectionConfigurer.Name("\t \n") with an invalid name`,
+				`*fixtures.ProjectionMessageHandler.Configure() called ProjectionConfigurer.Identity() with an invalid name "\t \n"`,
 				func(c dogma.ProjectionConfigurer) {
-					c.Name("\t \n")
+					c.Identity("\t \n", "<key>")
+					c.ConsumesEventType(fixtures.MessageA{})
+				},
+			),
+			Entry(
+				"when the handler configures an invalid key",
+				`*fixtures.ProjectionMessageHandler.Configure() called ProjectionConfigurer.Identity() with an invalid key "\t \n"`,
+				func(c dogma.ProjectionConfigurer) {
+					c.Identity("<name>", "\t \n")
 					c.ConsumesEventType(fixtures.MessageA{})
 				},
 			),
@@ -118,14 +128,14 @@ var _ = Describe("type ProjectionConfig", func() {
 				"when the handler does not configure any consumed event types",
 				`*fixtures.ProjectionMessageHandler.Configure() did not call ProjectionConfigurer.ConsumesEventType()`,
 				func(c dogma.ProjectionConfigurer) {
-					c.Name("<name>")
+					c.Identity("<name>", "<key>")
 				},
 			),
 			Entry(
 				"when the handler configures the same consumed event type multiple times",
 				`*fixtures.ProjectionMessageHandler.Configure() has already called ProjectionConfigurer.ConsumesEventType(fixtures.MessageA)`,
 				func(c dogma.ProjectionConfigurer) {
-					c.Name("<name>")
+					c.Identity("<name>", "<key>")
 					c.ConsumesEventType(fixtures.MessageA{})
 					c.ConsumesEventType(fixtures.MessageA{})
 				},